remove sanitisations that convert phoneNumbers to Ints

+++ src/device-registry/routes/api-v1.js
@@ -185,8 +185,7 @@ router.put(
           return isValid;
         })
         .withMessage("phoneNumber must be a valid one")
-        .bail()
-        .toInt(),
+        .bail(),
       body("height")
         .if(body("height").exists())
         .notEmpty()
@@ -462,8 +461,7 @@ router.post(
           return isValid;
         })
         .withMessage("phoneNumber must be a valid one")
-        .bail()
-        .toInt(),
+        .bail(),
       body("height")
         .if(body("height").exists())
         .notEmpty()
@@ -708,8 +706,7 @@ router.put(
           return isValid;
         })
         .withMessage("phoneNumber must be a valid one")
-        .bail()
-        .toInt(),
+        .bail(),
       body("height")
         .if(body("height").exists())
         .notEmpty()
@@ -1045,8 +1042,7 @@ router.put(
           return isValid;
         })
         .withMessage("phoneNumber must be a valid one")
-        .bail()
-        .toInt(),
+        .bail(),
       body("height")
         .if(body("height").exists())
         .notEmpty()
